fixes and tweaks for party map chats
